Split up the ValidationContext into a ChangingContext so that a custom ValidationContext can be created outside of Innovator.Client and also be able to re-use the new or existing code.
